backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic-settings==2.6.1
python-multipart==0.0.17
httpx==0.27.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
numpy==2.1.3
pillow==11.0.0
python-dotenv==1.0.1
